Buncha fixes
