+++ service/src/scripts/update-profile.ts
@@ -6,7 +6,7 @@ import { RPC_EXCEPTION } from '../utils/exit-codes'
 import { archLogger } from '../utils/chalk-theme'
 import { parseUpdateArgs } from '../utils/cli_parsers/parseUpdateArgs'
 import { ethers } from 'ethers'
-import { requestApproval } from './approve'
+import { requestApproval } from './approve_utils'
 
 validateEnvVars();
 
@@ -25,6 +25,7 @@ let freeBondDeposit = ethers.constants.Zero;
 
 if (freeBond.gt(ethers.constants.Zero)) {
   const approved = await requestApproval(
+    web3Interface,
     "You will need to approve Sarcophagus contracts to use your SARCO in order to deposit free bond.\nEnter 'approve' to authorize this, or else hit <ENTER> to continue without a deposit:"
   )
